gave file its own name seperate from reg seed file
